Refactor: 이름 코딩컨벤션에 맞게 수정 및 쿼리생성함수 분리
